[QMS-337] Reorganize GeoMath
