+++ src/qmapshack/mouse/CMouseRuler.cpp
@@ -25,7 +25,7 @@
 #include "gis/rte/CGisItemRte.h"
 #include "gis/trk/CGisItemTrk.h"
 #include "gis/wpt/CGisItemWpt.h"
-#include "GeoMath.h"
+#include "gis/GeoMath.h"
 #include "helpers/CDraw.h"
 #include "mouse/CMouseRuler.h"
 #include "mouse/CScrOptRuler.h"
