+++ src/qmapshack/mouse/line/ILineOp.cpp
@@ -21,12 +21,191 @@
 #include "CMainWindow.h"
 #include "gis/CGisDraw.h"
 #include "gis/CGisWorkspace.h"
+#include "gis/GeoMath.h"
 #include "gis/rte/router/CRouterSetup.h"
-#include "GeoMath.h"
 #include "mouse/line/ILineOp.h"
 #include "mouse/line/IMouseEditLine.h"
 
 #include <QtWidgets>
+
+struct segment_t
+{
+    segment_t() : idx11(NOIDX), idx12(NOIDX), idx21(NOIDX){}
+
+    void apply(const QPolygonF& coords, const QPolygonF& pixel, QPolygonF& segCoord, QPolygonF& segPixel, IDrawContext * context)
+    {
+        QPointF pt1 = px1;
+        QPointF pt2 = px2;
+
+        context->convertPx2Rad(pt1);
+        context->convertPx2Rad(pt2);
+
+        if(idx11 != NOIDX && idx21 != NOIDX)
+        {
+            if(idx12 == idx21)
+            {
+                segPixel.push_back(pixel[idx12]);
+                segCoord.push_back(coords[idx12]);
+            }
+            else if(idx11 < idx21)
+            {
+                for(int i = idx12; i <= idx21; i++)
+                {
+                    segPixel.push_back(pixel[i]);
+                    segCoord.push_back(coords[i]);
+                }
+            }
+            else if(idx11 > idx21)
+            {
+                for(int i = idx11; i > idx21; i--)
+                {
+                    segPixel.push_back(pixel[i]);
+                    segCoord.push_back(coords[i]);
+                }
+            }
+        }
+    }
+
+
+    qint32 idx11;
+    qint32 idx12;
+    qint32 idx21;
+
+    QPointF px1;
+    QPointF px2;
+};
+
+static inline qreal distance(const QPointF &pa, const QPointF &pb)
+{
+    const qreal &dx = pa.x() - pb.x();
+    const qreal &dy = pa.y() - pb.y();
+    return qSqrt(dx * dx + dy * dy);
+}
+
+void GPS_Math_SubPolyline(const QPointF& pt1, const QPointF& pt2, qint32 threshold, const QPolygonF& pixel, segment_t &result)
+{
+    projXY p1, p2;
+    qreal dx, dy;                // delta x and y defined by p1 and p2
+    qreal d_p1_p2;              // distance between p1 and p2
+    qreal x, y;                  // coord. (x,y) of the point on line defined by [p1,p2] close to pt
+    qreal shortest1 = threshold;
+    qreal shortest2 = threshold;
+    qint32 idx11 = NOIDX, idx21 = NOIDX, idx12 = NOIDX;
+
+    QPointF pt11;
+    QPointF pt21;
+
+    // find points on line closest to pt1 and pt2
+    const qint32 len = pixel.size();
+    for(qint32 i = 1; i < len; ++i)
+    {
+        p1.u = pixel[i - 1].x();
+        p1.v = pixel[i - 1].y();
+        p2.u = pixel[i].x();
+        p2.v = pixel[i].y();
+
+        dx = p2.u - p1.u;
+        dy = p2.v - p1.v;
+        d_p1_p2 = qSqrt(dx * dx + dy * dy);
+
+        // find point on line closest to pt1
+        // ratio u the tangent point will divide d_p1_p2
+        qreal u = ((pt1.x() - p1.u) * dx + (pt1.y() - p1.v) * dy) / (d_p1_p2 * d_p1_p2);
+
+        if(u >= 0.0 && u <= 1.0)
+        {
+            x = p1.u + u * dx;
+            y = p1.v + u * dy;
+
+            qreal distance = qSqrt((x - pt1.x()) * (x - pt1.x()) + (y - pt1.y()) * (y - pt1.y()));
+
+            if(distance < shortest1)
+            {
+                idx11 = i - 1;
+                idx12 = i;
+                pt11.setX(x);
+                pt11.setY(y);
+                shortest1 = distance;
+            }
+        }
+
+        // find point on line closest to pt2
+        // ratio u the tangent point will divide d_p1_p2
+        u = ((pt2.x() - p1.u) * dx + (pt2.y() - p1.v) * dy) / (d_p1_p2 * d_p1_p2);
+
+        if(u >= 0.0 && u <= 1.0)
+        {
+            x = p1.u + u * dx;
+            y = p1.v + u * dy;
+
+            qreal distance = qSqrt((x - pt2.x()) * (x - pt2.x()) + (y - pt2.y()) * (y - pt2.y()));
+
+            if(distance < shortest2)
+            {
+                idx21 = i - 1;
+                pt21.setX(x);
+                pt21.setY(y);
+                shortest2 = distance;
+            }
+        }
+    }
+
+    // if 1st point can't be found test for distance to both ends
+    if(idx11 == NOIDX)
+    {
+        QPointF px = pixel.first();
+        qreal dist = distance(px, pt1);
+        if(dist < (threshold << 1))
+        {
+            idx11 = 0;
+            idx12 = 1;
+            pt11 = px;
+        }
+        else
+        {
+            px = pixel.last();
+            qreal dist = distance(px, pt1);
+            if(dist < (threshold << 1))
+            {
+                idx11 = pixel.size() - 2;
+                idx12 = pixel.size() - 1;
+                pt11 = px;
+            }
+        }
+    }
+
+    // if 2nd point can't be found test for distance to both ends
+    if(idx21 == NOIDX)
+    {
+        QPointF px = pixel.first();
+        qreal dist = distance(px, pt2);
+
+        if(dist < (threshold << 1))
+        {
+            idx21 = 0;
+            pt21 = px;
+        }
+        else
+        {
+            px = pixel.last();
+            qreal dist = distance(px, pt2);
+            if(dist < (threshold << 1))
+            {
+                idx21 = pixel.size() - 2;
+                pt21 = px;
+            }
+        }
+    }
+
+    //    qDebug() << pixel.size() << idx11 << idx12 << idx21 << pt1 << pt2 << pt11 << pt21;
+
+    result.idx11 = idx11;
+    result.idx12 = idx12;
+    result.idx21 = idx21;
+    result.px1   = pt11;
+    result.px2   = pt21;
+}
+
 
 ILineOp::ILineOp(SGisLine& points, CGisDraw *gis, CCanvas *canvas, IMouseEditLine *parent)
     : QObject(parent)
